refact(project) debug functions

+++ js/fonctions.js
@@ -7,7 +7,7 @@ const selectionneQuantité = () => {
         quantiteArticle.innerText += i;
         quantiteArticle.append(choixQuantité);
     }
-}
+};
 
 
 // Choix de la couleur du nounours
@@ -18,14 +18,23 @@ const couleurArticle = (data) => {
         couleurChoisie.innerText = data.colors[i];
         couleur.append(couleurChoisie);
     }
+};
+
+const btnAllerPanier = document.querySelector(".btnPagePanier");
+if(btnAllerPanier){
+    btnAllerPanier.addEventListener('click',()=>{
+        document.location.href = './panier.html';
+});
 }
+
+
 
 // PANIER
 
 // Sauvegarder un article dans localStorage
 let sauvegarderArticle = () =>{
     localStorage.setItem('article',JSON.stringify(mesArticles));
-}
+};
 
 // Charger le panier
 let chargerPanier = () =>{
@@ -33,7 +42,7 @@ let chargerPanier = () =>{
 }
 if(localStorage.getItem('article') != null){
 chargerPanier();
-}
+};
 
 // Bouton incrémenter article
 const bntIncrementerQuantiter = (id, couleur) =>{
@@ -44,7 +53,7 @@ const bntIncrementerQuantiter = (id, couleur) =>{
     }
     sauvegarderArticle();
     location.reload();
-}
+};
 
 // Bouton décrementer compteur article
 const bntDecrementerQuantiter = (id, couleur) =>{
@@ -55,29 +64,39 @@ const bntDecrementerQuantiter = (id, couleur) =>{
     }
     sauvegarderArticle();
     location.reload();
-}
+};
 
 //Total panier
 const totalPanier = () => {
     let totalCommande = 0;
     for(article of mesArticles){
-        totalCommande += (article.prix / 200) * article.quantite
+        totalCommande += (article.prix / 200) * article.quantite;
     }
     let totalArticle = document.querySelector("#TotalPanier");
     return totalArticle.innerHTML = "Total de votre Commande : " +  totalCommande.toFixed(2) + " € ";
-}
+};
 
 // Bouton Supprimer un article du panier
 const btnSupprimerArticlePanier = () => {
     mesArticles.splice(article, 1);
     sauvegarderArticle();
     location.reload();
-}
+};
 
 // Bouton Supprimer le panier
 let supprimerPanier = document.querySelector('.suppPanier');
+if(supprimerPanier){
     supprimerPanier.addEventListener('click', (event)=>{
         localStorage.clear();
         location.reload();
+    });
+}
+
+
+const continuerAchat = document.querySelector(".btnContinuerAchat");
+if(continuerAchat){
+    continuerAchat.addEventListener('click',()=>{
+        document.location.href = './index.html';
     })
+}
